requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0

# AI and Machine Learning
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.24.4
scipy>=1.11.4
joblib>=1.3.2
xgboost>=2.0.2
lightgbm>=4.1.0
shap>=0.43.0

# CrewAI and AI Automation
crewai>=0.28.0
langchain>=0.1.0
openai>=1.0.0

# Geospatial and mapping
geopandas>=0.14.1
folium>=0.15.1
geopy>=2.4.1
shapely>=2.0.2
rasterio>=1.3.9

# Data processing and APIs
requests>=2.31.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
selenium>=4.16.0

# Scheduling and background tasks
celery>=5.3.4
redis>=5.0.1
apscheduler>=3.10.4

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
typer>=0.9.0
httpx>=0.25.2

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0

# Original pygame (keeping for reference)
pygame>=2.0.0